Update Telegram.Bot nuget package version

+++ src/Commands/IBotCommand.cs
@@ -9,6 +9,6 @@ namespace Soulfire.Bot.Commands
         string Description { get; }
         bool InternalCommand { get; }
 
-        Task Execute(IChatService chatService, long chatId, int userId, int messageId, string? commandText);
+        Task Execute(IChatService chatService, long chatId, long userId, int messageId, string? commandText);
     }
 }
