+++ src/Commands/PingCommand.cs
@@ -10,7 +10,7 @@ namespace Soulfire.Bot.Commands
 
         public bool InternalCommand => false;
 
-        public async Task Execute(IChatService chatService, long chatId, int userId, int messageId, string? commandText)
+        public async Task Execute(IChatService chatService, long chatId, long userId, int messageId, string? commandText)
         {
             await chatService.SendMessage(chatId, "pong");
         }
